+++ src/words.hpp
@@ -1,0 +1,29 @@
+//
+// Semigroups package for GAP
+// Copyright (C) 2023 James D. Mitchell
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+//
+
+#ifndef SEMIGROUPS_SRC_WORDS_HPP_
+#define SEMIGROUPS_SRC_WORDS_HPP_
+
+// Forward decl
+namespace gapbind14 {
+  class Module;
+}  // namespace gapbind14
+
+void init_words(gapbind14::Module&);
+
+#endif  // SEMIGROUPS_SRC_WORDS_HPP_
